requirements.txt
telebot~=0.0.5
pyTelegramBotAPI~=4.25.0
pillow~=11.0.0
python-dotenv~=1.0.1